remove old PFBlockAlgo, all remaining discrepancies tracked down to order dependencies
