+++ RecoParticleFlow/PFProducer/plugins/PFBlockProducer.h
@@ -41,57 +41,8 @@ class PFBlockProducer : public edm::EDProducer {
   virtual void produce(edm::Event&, const edm::EventSetup&) override;
 
  private:
-
-  
-
-  edm::InputTag   inputTagRecTracks_;
-  edm::InputTag   inputTagGsfRecTracks_;
-  edm::InputTag   inputTagConvBremGsfRecTracks_;
-  edm::InputTag   inputTagRecMuons_;
-  edm::InputTag   inputTagPFNuclear_;
-  edm::InputTag   inputTagPFClustersECAL_;
-  edm::InputTag   inputTagPFClustersHCAL_;
-  edm::InputTag   inputTagPFClustersHO_;
-  edm::InputTag   inputTagPFClustersHFEM_;
-  edm::InputTag   inputTagPFClustersHFHAD_;
-  edm::InputTag   inputTagPFClustersPS_;
-  edm::InputTag   inputTagPFConversions_;
-  edm::InputTag   inputTagPFV0_;
-  edm::InputTag   inputTagEGPhotons_;
-  edm::InputTag   inputTagSCBarrel_;
-  edm::InputTag   inputTagSCEndcap_;  
-  
-  // Link track and HCAL clusters to HO clusters ?
-  bool useHO_;
-
   /// verbose ?
   bool   verbose_;
-
-  /// use NuclearInteractions ?
-  bool   useNuclear_;
-
-  /// use EG photons ? 
-  bool useEGPhotons_;
-  
-  /// use SuperClusters ? 
-  bool useSuperClusters_;  
-  
-  /// switch on/off Conversions
-  bool  useConversions_;  
-  
-  /// switch on/off Conversions Brem Recovery
-  bool   useConvBremGsfTracks_;
-
-  /// switch on/off V0
-  bool useV0_;
-
-  /// Particle Flow at HLT ?
-  bool usePFatHLT_;
-
-  // Glowinski & Gouzevitch
-  // Use the optimized KDTree Track/Ecal linker?
-  bool useKDTreeTrackEcalLinker_;
-  // !Glowinski & Gouzevitch
 
   /// Particle flow block algorithm 
   PFBlockAlgo            pfBlockAlgo_;
