+++ RecoParticleFlow/PFProducer/python/particleFlowBlock_cff.py
@@ -6,5 +6,3 @@ import FWCore.ParameterSet.Config as cms
 # are not stored in the event!
 # include "RecoTracker/TrackProducer/data/CTFFinalFitWithMaterial.cff"
 from RecoParticleFlow.PFProducer.particleFlowBlock_cfi import *
-from RecoParticleFlow.PFProducer.particleFlowBlockNew_cfi import *
-
